ecop etpy diffs dist cmpr and flags
- Done with implementing objective funtions using distributions.

- A flag to turn this feature one off is also added.

- In prepare.py, a global flag 'extrapolate_flag' is added. When True,
uses the extrapolation feature in all interp1d objects that are used
here. The problem is that with extrapolation the distribution fit is
better but the empirical copulas and properties that are indirectly
affected by the objective function (e.g. asymm_2 dist fitting affects
scorr and etpy) lose quality as well.